requirements.txt
streamlit
pandas
numpy
pillow
tensorflow-cpu
scikit-learn